requirements.txt
flask==3.0.0
werkzeug==3.0.1
langchain>=0.1.10
langchain-community>=0.0.28
langchain-core>=0.1.10
langchain-groq>=0.2.1
langchain-nomic
langchain-chroma
python-dotenv==1.0.1
PyPDF2
chromadb
tiktoken
PyMuPDF
Pillow
numpy
pypdf
pdfminer.six
python-magic-bin; sys_platform == 'win32'
rapidocr-onnxruntime
langchain-experimental
python-poppler; sys_platform != 'win32'
